requirements.txt
pydantic-ai==0.1.3
pydantic-ai-slim[mcp]==0.1.3
pydantic-ai-slim[groq]==0.1.3
mcp-server-fetch==2025.4.7
fastapi
uvicorn